pyproject.toml
[tool.poetry]
name = "wallet-drf-api"
version = "0.1.0"
description = ""
authors = ["khalaimovda <halaimov_15@mail.ru>"]
readme = "README.md"
packages = [{include = "wallet_drf_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
alembic = "^1.13.1"
sqlalchemy = "^2.0.29"
pymysql = "^1.1.0"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"
django-filter = "^24.2"
mysqlclient = "^2.2.4"
cryptography = "^42.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
